Minor renaming, restructuring, bug fixes, and file clean-up.
