+++ landsites_tools/inputprep/__init__.py
@@ -1,1 +1,3 @@
+#!/usr/bin/env python
+
 """Sub-package to prepare input data."""
