+++ landsites_tools/postprocess/postprocess.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+
+"""postprocess.py: post-process the CTSM simulation results, e.g. create plots
+for output variables defined in a settings file."""
