+++ landsites_tools/run.py
@@ -1,1 +1,3 @@
+#!/usr/bin/env python
+
 ### Main runner function will be added here
