+++ landsites_tools/simulation/__init__.py
@@ -1,1 +1,3 @@
+#!/usr/bin/env python
+
 """Sub-package to prepare and execute simulations."""
