+++ landsites_tools/simulation/run_cases.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+
+"""run_cases.py: run CTSM simulations, requires existing case directories."""
+
+### Coming soon
